refactor authorizer
